+++ tests/unit/store/test_parquet.py
@@ -5,34 +5,59 @@ from pytest_lazyfixture import lazy_fixture
 from blueetl.store import parquet as test_module
 
 
-# fastparquet 0.8.1 fails to write DataFrames with MultiIndexes without names,
-# but probably it's not a good idea to use them anyway. See the code at:
-# https://github.com/dask/fastparquet/blob/34069fe2a41a7491e5b7b1f1b2cae9c41176f7b8/fastparquet/util.py#L140-L144
-
-
 @pytest.mark.parametrize(
     "df",
     [
         lazy_fixture("storable_df_with_unnamed_index"),
         lazy_fixture("storable_df_with_named_index"),
-        # lazy_fixture("storable_df_with_unnamed_multiindex"),  # failing with fastparquet
         lazy_fixture("storable_df_with_named_multiindex"),
+        # fastparquet 0.8.1 fails to write DataFrames with MultiIndexes without names,
+        # but probably it's not a good idea to use them anyway. See the code at:
+        # https://github.com/dask/fastparquet/blob/34069fe2a41a7491e5b7b1f1b2cae9c41176f7b8/fastparquet/util.py#L140-L144
+        # lazy_fixture("storable_df_with_unnamed_multiindex"),
     ],
 )
 @pytest.mark.parametrize(
-    "dump_engine, load_engine",
+    "dump_options, load_options",
     [
-        ("pyarrow", "pyarrow"),
-        ("fastparquet", "fastparquet"),
-        # ("pyarrow", "fastparquet"),  # error: column e is loaded as float64 instead of object
-        # ("fastparquet", "pyarrow"),  # error: column h is loaded as bytes instead of object
+        # test the configuration actually used
+        (None, None),
+        # test other possible configurations not used yet
+        ({"engine": "pyarrow", "index": None}, {"engine": "pyarrow"}),
+        ({"engine": "fastparquet", "index": None}, {"engine": "fastparquet"}),
+        pytest.param(
+            {"engine": "fastparquet", "index": True},
+            {"engine": "fastparquet"},
+            marks=pytest.mark.xfail(
+                reason="Fails because index names are different ('index', None)",
+                raises=AssertionError,
+            ),
+        ),
+        pytest.param(
+            {"engine": "pyarrow"},
+            {"engine": "fastparquet"},
+            marks=pytest.mark.xfail(
+                reason="Fails because column e is loaded as float64 instead of object",
+                raises=AssertionError,
+            ),
+        ),
+        pytest.param(
+            {"engine": "fastparquet"},
+            {"engine": "pyarrow"},
+            marks=pytest.mark.xfail(
+                reason="Fails because column h is loaded as bytes instead of object",
+                raises=AssertionError,
+            ),
+        ),
     ],
 )
-def test_dump_load_roundtrip(tmp_path, df, dump_engine, load_engine):
+def test_dump_load_roundtrip(tmp_path, df, dump_options, load_options):
     name = "myname"
     store = test_module.ParquetStore(tmp_path)
-    store._dump_options["engine"] = dump_engine
-    store._load_options["engine"] = load_engine
+    if dump_options is not None:
+        store._dump_options = dump_options
+    if load_options is not None:
+        store._load_options = load_options
 
     store.dump(df, name)
     result = store.load(name)